agent/requirements.txt
requests
pyserial
